LayoutTests/fast/frames/seamless/seamless-sandbox-flag-expected.txt

PASS iframe.seamless is true
PASS nestedFrame.seamless is true
PASS window.getComputedStyle(iframe).width is "200px"
FAIL window.getComputedStyle(iframe).height should be 150px. Was 0px.
